refactor(testsupport): build the in-repo helpers on Rask.Testing's public API
Rask.TestSupport.Markup.Attr and Rask.Testing's MarkupQuery.Attr were the same algorithm, character
for character, comments included — two copies of one scanner, already at risk of drifting apart. The
Markup doc comment says it exists to consolidate per-file copies "that had drifted"; keeping a second
one invites exactly that again, one layer up.

TestSupport now references Rask.Testing and the duplicate is deleted. Markup.Attr/Attrs come from the
shipped package, which puts the public API on the compile path of all seven suites that use
TestSupport — the point of dogfooding it.

What stays is what the package deliberately doesn't ship, and TestSupport's charter is now sayable in
one line: the Assert-calling helpers (Rask.Testing is test-framework-agnostic and must stay so) and
the live-wire-payload helpers (below the HTML + handler-dispatch seam the package covers). They move
to MarkupAssert — two classes named Markup in scope together make every unqualified use ambiguous
(CS0104), which is not theoretical: it is what the compiler said when this was tried the other way.

RenderHarness.EmptyServices stays alongside RaskTest's default provider. Not duplication — one builds
a real provider, the other resolves nothing on purpose so the package takes no DI dependency.

Naming Rask.Testing's type `Markup` was the choice that made this cheap: all 136 Markup.Attr call
sites keep compiling untouched and only the 75 Assert-coupled ones move. Zero behaviour change — the
rename is compiler-verified, no Assert line in the diff, and the full unit gate is green.
