+++ tests/Rask.Bootstrap.Tests/Rask.Bootstrap.Tests.csproj
@@ -30,6 +30,7 @@
     <Using Include="Rask.Core.Components.Generated" Static="true"/>
     <Using Include="Rask.Bootstrap"/>
     <Using Include="Rask.Bootstrap.Generated" Static="true"/>
+    <Using Include="Rask.Testing"/>
     <Using Include="Rask.TestSupport"/>
   </ItemGroup>
