+++ tests/Rask.Native.Tests/Rask.Native.Tests.csproj
@@ -23,6 +23,7 @@
 
   <ItemGroup>
     <Using Include="Xunit"/>
+    <Using Include="Rask.Testing"/>
     <Using Include="Rask.TestSupport"/>
     <Using Include="Rask.Native.Tests.Infrastructure.NativeSessionHarness" Static="true"/>
   </ItemGroup>
