+++ tests/Rask.Server.Tests/Rask.Server.Tests.csproj
@@ -30,6 +30,7 @@
     <Using Include="Xunit"/>
     <Using Include="Rask.Core.Components.Generated" Static="true"/>
     <Using Include="Rask.Core.Routing.Generated" Static="true"/>
+    <Using Include="Rask.Testing"/>
     <Using Include="Rask.TestSupport"/>
   </ItemGroup>
