+++ tests/Rask.Server.Tests/Security/WebSocketFrameSizeTests.cs
@@ -55,7 +55,7 @@ public class WebSocketFrameSizeTests
         // fine under the default cap.
         using var host = RaskTestHost.Create<TestApp>();
         var get = await host.Http.GetAsync("/");
-        var sessionId = Markup.SessionId(await get.Content.ReadAsStringAsync());
+        var sessionId = MarkupAssert.SessionId(await get.Content.ReadAsStringAsync());
 
         using var ws = await host.WebSockets.ConnectAsync(host.WebSocketUri, CancellationToken.None);
         await ws.SendJsonAsync(new { type = "hello", session = sessionId });
