+++ tests/Rask.Server.Tests/WebSockets/InitialRenderDiffTests.cs
@@ -20,8 +20,8 @@ public class InitialRenderDiffTests
     {
         using var host = RaskTestHost.Create<TestApp>();
         var initialHtml = await (await host.Http.GetAsync("/start")).Content.ReadAsStringAsync();
-        var sessionId = Markup.SessionId(initialHtml);
-        var handlerId = Markup.FirstHandlerId(initialHtml);
+        var sessionId = MarkupAssert.SessionId(initialHtml);
+        var handlerId = MarkupAssert.FirstHandlerId(initialHtml);
 
         using var ws = await host.WebSockets.ConnectAsync(host.WebSocketUri, CancellationToken.None);
         await ws.SendJsonAsync(new { type = "hello", session = sessionId });
