+++ tests/Rask.Server.Tests/WebSockets/PayloadDedupTests.cs
@@ -10,8 +10,8 @@ public class PayloadDedupTests
         using var host = RaskTestHost.Create<NoOpApp>();
         var initial = await host.Http.GetAsync("/start");
         var initialHtml = await initial.Content.ReadAsStringAsync();
-        var sessionId = Markup.SessionId(initialHtml);
-        var handlerId = Markup.FirstHandlerId(initialHtml);
+        var sessionId = MarkupAssert.SessionId(initialHtml);
+        var handlerId = MarkupAssert.FirstHandlerId(initialHtml);
 
         using var ws = await host.WebSockets.ConnectAsync(host.WebSocketUri, CancellationToken.None);
         await ws.SendJsonAsync(new { type = "hello", session = sessionId });
