+++ tests/Rask.Validation.DataAnnotations.Tests/Rask.Validation.DataAnnotations.Tests.csproj
@@ -30,6 +30,7 @@
     <Using Include="Rask.Core.Components.Generated" Static="true"/>
     <Using Include="Rask.Validation.DataAnnotations"/>
     <Using Include="Rask.Validation.DataAnnotations.Generated" Static="true"/>
+    <Using Include="Rask.Testing"/>
     <Using Include="Rask.TestSupport"/>
     <Using Include="Rask.Validation.DataAnnotations.Tests.ValidationTestSupport" Static="true"/>
   </ItemGroup>
